requirements.txt
python-telegram-bot==21.10
yandex_cloud_ml_sdk==0.2.4
